[src] Create base class of collator for BirdCLEF
